backend/requirements.txt
alabaster==1.0.0
argcomplete==3.6.2
asgiref==3.9.1
babel==2.17.0
# bashfuscator @ file:///home/teddybear/bigpapi_project/Bashfuscator
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
django-widget-tweaks==1.5.0
docopt==0.6.2
docutils==0.21.2
gunicorn==23.0.0
idna==3.10
imagesize==1.4.1
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
pipreqs==0.4.13
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
pyperclip==1.9.0
pytest==8.4.1
python-decouple==3.8
python-dotenv==1.1.1
requests==2.32.4
roman-numerals-py==3.1.0
setuptools==80.9.0
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.3
stripe==12.3.0
typing_extensions==4.14.1
urllib3==2.5.0
yarg==0.1.10
djangorestframework
